requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
binary==1.0.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
dataclasses-json==0.6.4
Django==5.0.3
djangorestframework==3.15.1
filelock==3.13.3
frozenlist==1.4.1
fsspec==2024.3.1
google-ai-generativelanguage==0.4.0
google-api-core==2.18.0
google-auth==2.29.0
google-generativeai==0.4.1
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.1
grpcio-status==1.62.1
huggingface-hub==0.22.2
idna==3.6
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.14
langchain-community==0.0.30
langchain-core==0.1.37
langchain-text-splitters==0.0.1
langsmith==0.1.38
marshmallow==3.21.1
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
orjson==3.10.0
packaging==23.2
pgvector==0.2.5
proto-plus==1.23.0
protobuf==4.25.3
psycopg==3.1.18
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.6.4
pydantic_core==2.16.3
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rsa==4.9
safetensors==0.4.2
SQLAlchemy==2.0.29
sqlparse==0.4.4
tenacity==8.2.3
tokenizers==0.15.2
tqdm==4.66.2
transformers==4.39.2
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
yarl==1.9.4
